cln: renaming func
